+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 New features:
 
 * Updated to Ruby 3.2.2 (#3039, @eregon, @andrykonchin).
+* TruffleRuby Native on Oracle GraalVM on Linux now uses the G1 garbage collector which is much faster.
 
 Bug fixes: